amos: Change to incremental burst type
It seems that the most widely supported burst types are INCR and WRAP bursts, at least on
the Xilinx infrastructure. I am pretty confident that most of the infrastructure should have
INCR bursts implemented since this is one of the main benefits of having an AXI-based
infrastructure.

In particular this change fixes the burst type to incremental of the split RMW sequence, essentially
mitigating an issue of the Xilinx SmartConnect blocks which do not support FIXED bursts.