Fix non blocking assignment in lrsc FF
The blocking assignment causes the `clr_len_q` to taken an incorrect value during AXI bursts. This leads to an underflow of the length value and the counter has to decrement until it reaches one and the `AW_BURST` state can finally be left. This issue was discovered under Verilator. The fix results in a 10x simulation speed improvement.