Add support for limited number of reservations
If limited LR/SC support is OK, NUM_RESERVATIONS can be set to reduce
how many active atomics are supported, significantly improving PPA. This
means that some reservations may be cleared without another actor
accessing that address, as too many actors are reserving at the same
time. If the parameter is not set, existing code will remain, so is
backwards-compatible.